budget

+++ src/app/components/login/login.component.ts
@@ -35,9 +35,12 @@ export class LoginComponent {
         response =>{
           console.log(response);
           if(response.data != undefined){
-            localStorage.clear();
-            localStorage.setItem('user',JSON.stringify(response.data));
-            localStorage.setItem('token', response.jwt);
+            if (typeof window !== 'undefined') {
+              localStorage.clear();
+              localStorage.setItem('user', JSON.stringify(response.data));
+              localStorage.setItem('token', response.jwt);
+            }
+          
             this._router.navigate(['/dashboard']);
 
           } else {
